SEM15LUISCALLE: Restructuracion de UsuarioRepository
