+++ src/main/java/com/transport/management/config/ApplicationConfig.java
@@ -1,6 +1,5 @@
 package com.transport.management.config;
 
-import com.transport.management.repositories.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,6 +12,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import com.transport.management.modules.usuario.UsuarioRepository;
 
 @Configuration
 public class ApplicationConfig {
