+++ src/main/java/com/transport/management/modules/usuario/UsuarioRepository.java
@@ -1,6 +1,5 @@
-package com.transport.management.repositories;
+package com.transport.management.modules.usuario;
 
-import com.transport.management.modules.usuario.UsuarioEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 import org.springframework.stereotype.Repository;
